XMLHttpRequest/open-user-password-non-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - user/pass argument and non same-origin URL</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var m = "GET",
          u = "http://test2.w3.org/",
          a = false
        test(function() {
          var client = new XMLHttpRequest()
          client.open(m, u, a, "x")
          assert_equals(client.readyState, 1, "open() was successful - 1")
          var client2 = new XMLHttpRequest()
          client2.open(m, u, a, "x", "x")
          assert_equals(client2.readyState, 1, "open() was successful - 2")
        })
    </script>
  </body>
</html>
